requirements.txt
numpy
matplotlib
camel-ai
pyyaml
pandas
python-dotenv
requests
hypernetx
fastjsonschema